Remove "coboleditor.enable_auto_tasks" as it was incomplete and did work as expected. Ensure the use of  VSQuickCOBOLParse.getCachedObject() checks for undefined, as needed
Simplify, copybook search by removing loop of workspace folders and
include them in the copybook folders during extension init.

Incldue workspace folder event, to pick up folder name changes,
as the dirs are cached in the copybook folders list, without it
will be out of sync.

Only show COBOL log windows when the command is executed by the
user, not on startup.