Treat slash in col 7 as comment line
Added code to treat any line with a '/' or '*' in col 7 to be highlighted as comment line.  The '/' is a printer control character but some shops will add text to the line.
Also, added code to treat any line with '*' in col 7 to be highlighted.  Currently, if you have a special character such as @ in any column 1-6, the line will not be highlighted as a comment line.  Some shops use col 1-6 to indicate changed lines with a release tag containing special characters.  This change will ignore any character in 1-6 as long as col 7 has either an * or /.